Conflicts:
	modules/core/src/main/java/org/openlmis/core/domain/FacilityTypeApprovedProduct.java
	modules/core/src/main/java/org/openlmis/core/domain/RequisitionGroup.java
	modules/core/src/main/java/org/openlmis/core/domain/Right.java
	modules/core/src/main/java/org/openlmis/core/domain/RightType.java
	modules/core/src/main/java/org/openlmis/core/domain/User.java
	modules/core/src/main/java/org/openlmis/core/repository/FacilityApprovedProductRepository.java
	modules/core/src/main/java/org/openlmis/core/repository/FacilityRepository.java
	modules/core/src/main/java/org/openlmis/core/repository/GeographicZoneRepository.java
	modules/core/src/main/java/org/openlmis/core/repository/ProductRepository.java
	modules/core/src/main/java/org/openlmis/core/repository/ProgramProductRepository.java
	modules/core/src/main/java/org/openlmis/core/repository/RequisitionGroupProgramScheduleRepository.java
	modules/core/src/main/java/org/openlmis/core/repository/RequisitionGroupRepository.java
	modules/core/src/main/java/org/openlmis/core/repository/SupplyLineRepository.java
	modules/core/src/main/java/org/openlmis/core/repository/mapper/FacilityApprovedProductMapper.java
	modules/core/src/main/java/org/openlmis/core/repository/mapper/FacilityMapper.java
	modules/core/src/main/java/org/openlmis/core/repository/mapper/GeographicLevelMapper.java
	modules/core/src/main/java/org/openlmis/core/repository/mapper/GeographicZoneMapper.java
	modules/core/src/main/java/org/openlmis/core/repository/mapper/ProcessingScheduleMapper.java
	modules/core/src/main/java/org/openlmis/core/repository/mapper/ProgramProductMapper.java
	modules/core/src/main/java/org/openlmis/core/repository/mapper/RequisitionGroupMapper.java
	modules/core/src/main/java/org/openlmis/core/repository/mapper/RequisitionGroupProgramScheduleMapper.java
	modules/core/src/main/java/org/openlmis/core/repository/mapper/SupervisoryNodeMapper.java
	modules/core/src/main/java/org/openlmis/core/repository/mapper/SupplyLineMapper.java
	modules/core/src/main/java/org/openlmis/core/repository/mapper/UserMapper.java
	modules/core/src/main/java/org/openlmis/core/service/FacilityService.java
	modules/core/src/main/java/org/openlmis/core/service/GeographicZoneService.java
	modules/core/src/main/java/org/openlmis/core/service/ProductCategoryService.java
	modules/core/src/main/java/org/openlmis/core/service/ProductService.java
	modules/core/src/main/java/org/openlmis/core/service/RequisitionGroupProgramScheduleService.java
	modules/core/src/main/java/org/openlmis/core/service/RequisitionGroupService.java
	modules/core/src/main/java/org/openlmis/core/service/SupervisoryNodeService.java
	modules/core/src/main/java/org/openlmis/core/service/SupplyLineService.java
	modules/core/src/main/java/org/openlmis/core/service/UserService.java
	modules/core/src/test/java/org/openlmis/core/repository/GeographicZoneRepositoryTest.java
	modules/core/src/test/java/org/openlmis/core/repository/mapper/GeographicLevelMapperIT.java
	modules/core/src/test/java/org/openlmis/core/service/RoleRightsServiceTest.java
	modules/core/src/test/java/org/openlmis/core/service/SupplyLineServiceTest.java
	modules/db/src/main/resources/seed/seed.sql
	modules/openlmis-web/src/main/java/org/openlmis/web/controller/FacilityApprovedProductController.java
	modules/openlmis-web/src/main/java/org/openlmis/web/controller/FacilityController.java
	modules/openlmis-web/src/main/java/org/openlmis/web/controller/FacilityProgramProductController.java
	modules/openlmis-web/src/main/java/org/openlmis/web/controller/GeographicZoneController.java
	modules/openlmis-web/src/main/java/org/openlmis/web/controller/OrderController.java
	modules/openlmis-web/src/main/java/org/openlmis/web/controller/PODController.java
	modules/openlmis-web/src/main/java/org/openlmis/web/controller/ProductController.java
	modules/openlmis-web/src/main/java/org/openlmis/web/controller/ProgramProductController.java
	modules/openlmis-web/src/main/java/org/openlmis/web/controller/RequisitionGroupController.java
	modules/openlmis-web/src/main/java/org/openlmis/web/controller/SupervisoryNodeController.java
	modules/openlmis-web/src/main/java/org/openlmis/web/controller/SupplyLineController.java
	modules/openlmis-web/src/main/java/org/openlmis/web/controller/UserController.java
	modules/openlmis-web/src/main/resources/messages_en.properties
	modules/openlmis-web/src/main/resources/messages_es.properties
	modules/openlmis-web/src/main/resources/messages_pt.properties
	modules/openlmis-web/src/main/webapp/public/js/facility/module/facility-module.js
	modules/openlmis-web/src/main/webapp/public/js/role/controller/role-controller.js
	modules/openlmis-web/src/main/webapp/public/js/shared/controller/login-controller.js
	modules/openlmis-web/src/main/webapp/public/js/shared/services/services.js
	modules/openlmis-web/src/main/webapp/public/js/user/controller/user-search-controller.js
	modules/openlmis-web/src/main/webapp/public/less/content.less
	modules/openlmis-web/src/main/webapp/public/pages/admin/facility/partials/create.html
	modules/openlmis-web/src/main/webapp/public/pages/admin/product/index.html
	modules/openlmis-web/src/main/webapp/public/pages/admin/product/partials/create.html
	modules/openlmis-web/src/main/webapp/public/pages/admin/requisition-group/index.html
	modules/openlmis-web/src/main/webapp/public/pages/admin/requisition-group/partials/create.html
	modules/openlmis-web/src/main/webapp/public/pages/admin/role/partials/create.html
	modules/openlmis-web/src/main/webapp/public/pages/admin/shared/manage-menu.html
	modules/openlmis-web/src/main/webapp/public/pages/admin/supervisory-node/index.html
	modules/openlmis-web/src/main/webapp/public/pages/admin/supervisory-node/partials/create.html
	modules/openlmis-web/src/main/webapp/public/pages/admin/user/partials/create.html
	modules/openlmis-web/src/main/webapp/public/pages/admin/user/partials/search.html
	modules/openlmis-web/src/main/webapp/public/pages/admin/user/partials/user-supervisory-role-assignment.html
	modules/openlmis-web/src/main/webapp/public/pages/logistics/rnr/partials/approve/approve.html
	modules/openlmis-web/src/main/webapp/public/pages/logistics/rnr/partials/create/index.html
	modules/openlmis-web/src/main/webapp/public/pages/logistics/rnr/partials/create/regimen-line-item.html
	modules/openlmis-web/src/main/webapp/public/pages/logistics/rnr/partials/create/rnr-full-supply.html
	modules/openlmis-web/src/main/webapp/public/pages/navigation.html
	modules/openlmis-web/src/test/java/org/openlmis/web/controller/PODControllerTest.java
	modules/openlmis-web/src/test/java/org/openlmis/web/controller/ProgramProductControllerTest.java
	modules/openlmis-web/src/test/java/org/openlmis/web/controller/RequisitionControllerTest.java
	modules/openlmis-web/src/test/java/org/openlmis/web/controller/RoleRightsControllerTest.java
	modules/openlmis-web/src/test/java/org/openlmis/web/controller/SupplyLineControllerTest.java
	modules/openlmis-web/src/test/java/org/openlmis/web/view/pdf/requisition/RequisitionPdfModelTest.java
	modules/openlmis-web/src/test/javascript/specs/user/controller/user-search-controller-test.js
	modules/requisition/src/main/java/org/openlmis/rnr/domain/RnrLineItem.java
	modules/requisition/src/main/java/org/openlmis/rnr/search/strategy/FacilityProgramDateRangeSearch.java
	modules/requisition/src/test/java/org/openlmis/rnr/domain/RnrLineItemTest.java
	modules/requisition/src/test/java/org/openlmis/rnr/service/RequisitionServiceTest.java
	modules/shipment/src/main/java/org/openlmis/order/service/OrderService.java
	modules/shipment/src/main/java/org/openlmis/shipment/repository/mapper/ShipmentMapper.java
	modules/shipment/src/test/java/org/openlmis/pod/service/PODServiceTest.java
	modules/shipment/src/test/java/org/openlmis/shipment/ShipmentLineItemTransformerTest.java

+++ modules/report/src/main/java/org/openlmis/report/service/NonReportingFacilityReportDataProvider.java
@@ -116,7 +116,7 @@ public class NonReportingFacilityReportDataProvider extends ReportDataProvider {
       header += "Program: " + this.reportMapper.getProgram(Integer.parseInt(program)).get(0).getName();
     }
     if (!zone.equals("0") && !zone.isEmpty() && !zone.endsWith("undefined")) {
-      header += "\nGeographic Zone: " + this.geographicZoneMapper.getById(Integer.parseInt(zone)).getName();
+      header += "\nGeographic Zone: " + this.geographicZoneMapper.getById(Long.parseLong(zone)).getName();
     }
 
     if (!facilityType.isEmpty() && !facilityType.equals("0") && !facilityType.endsWith("undefined")) {
